requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
pinecone-client>=3.0.0
python-dotenv
youtube-transcript-api
fastapi
uvicorn